changes for stats and admin panel

+++ LA_LIGA_REKREATIVO/Server/Models/League.cs
@@ -12,7 +12,7 @@
         public string? Logo { get; set; }
         public bool IsDeleted { get; set; }
         public bool IsActive { get; set; }
-        public double Coefficient { get; set; }
+        public decimal Coefficient { get; set; }
         public ICollection<Team> Teams { get; set; }
     }
 }
